api/pyproject.toml
[project]
name = "valor-api"
dynamic = ["version"]
description = "Backend for Valor evaluation store"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
    "fastapi[all]>=0.100.0",
    "PyJWT[crypto]",
    "GeoAlchemy2",
    "psycopg2-binary",
    "SQLAlchemy>=2.0",
    "Pillow >= 9.1.0",
    "numpy",
    "python-dotenv",
    "pydantic-settings",
    "structlog",
    "pgvector",
    "openai",
    "mistralai >= 1.0",
    "absl-py",
    "nltk",
    "rouge_score",
    "evaluate",
    "valor_lite",
]

[build-system]
requires = ["setuptools>=61.0", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
test = ["pytest", "coverage"]

[tool.black]
line-length = 79

[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.setuptools_scm]
root = ".."
